+++ src/main/java/com/escola/admin/repository/cliente/ClienteDependenteRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ClienteDependenteRepository extends CrudRepository<ClienteDependente, Integer> {
+public interface ClienteDependenteRepository extends CrudRepository<ClienteDependente, Long> {
 
     @Query("SELECT e FROM ClienteDependente e " +
             " WHERE e.cliente.id = :id")
-    Optional<List<ClienteDependente>> findAllByClienteId(Integer id);
+    Optional<List<ClienteDependente>> findAllByClienteId(Long id);
 }
